Adding functionality and handlers for auto events - Deal damage packet handles refactored to use new packet handler - Take damage packet handle refactored tot use new packet handler - Added a player disconnect handler for easier detection and handling - Auto event functionality added (but not tested) - Added ability to override monster stats

+++ src/net/server/channel/handlers/MoveLifeHandler.java
@@ -60,7 +60,7 @@ public final class MoveLifeHandler extends PacketHandler {
         moveId = slea.readShort();
         bSkill = slea.readByte();
         skill = slea.readByte();
-        skill_1 = slea.readInt() & 0xFF;
+        skill_1 = slea.readByte() & 0xFF;
         skill_2 = slea.readByte();
         skill_3 = slea.readByte();
         skill_4 = slea.readByte();
